Updated IUser to IPatient and fixed the update and delete crud functions

+++ frontend/firebase/IPsychiatrist.ts
@@ -29,19 +29,19 @@ const createPsychiatrist = async (psychiatrist: IPsychiatrist): Promise<void> =>
 
 /** To fetch a specific psychiatrist based on name;  
  * the below can technically do same but similar method currently in fetch data so wanted to stay consistent*/
-// const fetchPsychiatrist = async (firstName: string, lastName: string)=> {
-const fetchPsychiatrist = async (constraints: QueryConstraint[])=> {
+const fetchPsychiatrist = async (firstName: string, lastName: string)=> {
+// const fetchPsychiatrist = async (constraints: QueryConstraint[])=> {
     try {
         const baseQuery = collection(db, PYSCHIATRIST);
-        // const q = query(
-        //   baseQuery,
-        //   where("first_name", "==", firstName),
-        //   where("last_name", "==", lastName)
-        // );
         const q = query(
           baseQuery,
-          ...constraints
+          where("first_name", "==", firstName),
+          where("last_name", "==", lastName)
         );
+        // const q = query(
+        //   baseQuery,
+        //   ...constraints
+        // );
         
         
         const querySnapshot= await getDocs(q);
@@ -50,7 +50,7 @@ const fetchPsychiatrist = async (constraints: QueryConstraint[])=> {
           const psychiatrist = docData as IPsychiatrist;
           return psychiatrist;
         } else {
-          // throw new Error(`No psychiatrist found with the name: ${firstName} ${lastName}`);
+          throw new Error(`No psychiatrist found with the name: ${firstName} ${lastName}`);
         }
       } catch (error) {
         console.error("Error fetching professional data:", error);
@@ -77,20 +77,24 @@ const fetchAllPsychiatrist = async (constraints: QueryConstraint[]): Promise<IPs
     }
   }
 
-const updatePsychiatrist = async (id: string, updatedPsychiatrist: Partial<IPsychiatrist>): Promise<void> => {
+const updatePsychiatrist = async (id: string | undefined , updatedPsychiatrist: Partial<IPsychiatrist>): Promise<void> => {
   try {
-    const psychiatristRef = doc(db, PYSCHIATRIST, id);
+    if (id){
+      const psychiatristRef = doc(db, PYSCHIATRIST, id);
     await updateDoc(psychiatristRef, updatedPsychiatrist);
+    }
   } catch (error) {
     console.error("Error updating psychiatrist entry: ", error);
   }
   }
 
 
-  const deletePsychiatrist = async (id: string): Promise<void> => {
+  const deletePsychiatrist = async (id: string | undefined): Promise<void> => {
     try {
-      const psychiatristRef =doc(db, PYSCHIATRIST, id);
+      if (id){
+        const psychiatristRef =doc(db, PYSCHIATRIST, id);
         await deleteDoc(psychiatristRef);
+      }
     } catch (error) {
         console.error("Error deleting psychiatrist entry: ", error);
     }
